+++ src/installer/windows_vim_build_installer.ts
@@ -3,12 +3,14 @@ import * as path from "path";
 import * as core from "@actions/core";
 import {exec} from "@actions/exec";
 import * as io from "@actions/io";
+import {backportPatch} from "../patch";
 import {FixedVersion} from "../interfaces";
 import {VimBuildInstaller} from "./vim_build_installer";
 
 export class WindowsVimBuildInstaller extends VimBuildInstaller {
   async install(vimVersion: FixedVersion): Promise<void> {
     const reposPath = await this.cloneVim(vimVersion);
+    const extraArgs = await backportPatch(reposPath, vimVersion, this.isGUI);
     const arch = core.getInput("arch").includes("64") ? "x64" : "x86";
     const srcPath = path.join(reposPath, "src");
     const batPath = path.join(srcPath, "install.bat");
@@ -21,7 +23,7 @@ export class WindowsVimBuildInstaller extends VimBuildInstaller {
     rem Suppress progress animation
     sed -e "s/@<<$/@<< | sed -e 's#.*\\\\r.*##'/" Make_mvc.mak > Make_mvc2.mak
 
-    nmake -nologo -f Make_mvc2.mak ${guiOptions} FEATURES=HUGE IME=yes MBYTE=yes ICONV=yes DEBUG=no TERMINAL=yes
+    nmake -nologo -f Make_mvc2.mak ${guiOptions} FEATURES=HUGE IME=yes MBYTE=yes ICONV=yes DEBUG=no TERMINAL=yes ${extraArgs.join(" ")}
 
     copy /Y ..\\README.txt ..\\runtime
     copy /Y ..\\vimtutor.bat ..\\runtime